finish Student

+++ Library/Student.h
@@ -1,4 +1,6 @@
 #pragma once
+#include <iostream>
+#include <sstream>
 #include <vector>
 #include "Book.h"
 #include "Dao.h"
@@ -18,7 +20,7 @@ private:
 	int major;
 	char photo[200];
 	//一天一元计算公式
-	double money;
+	int money;
 	char email[30];
 	status_class status;
 public:
@@ -35,8 +37,8 @@ public:
 	int getMajor();
 	void setPhoto(string photo);
 	string getPhoto();
-	void setMoney(double money);
-	double getMoney();
+	void setMoney(int money);
+	int getMoney();
 	void setEmail(string email);
 	string getEmail();
 	void setStatus(status_class status);
